check unique and retweets

+++ readFiles.js
@@ -1,6 +1,47 @@
 const fs = require("fs");
 const brandConfig = require("./brandConfig.js");
 
+const tweetIsInArray = (newTweet, existingTweets) =>
+  existingTweets.find(
+    (existingTweet) => newTweet.tweetId === existingTweet.tweetId
+  );
+
+const returnOriginalTweet = (newTweet, existingTweets) => {
+  newTweet.retweetedId &&
+    existingTweets.find(
+      (existingTweet) => newTweet.retweetedId === existingTweet.tweetId
+    );
+};
+
+const renameFiles = () => {
+  const dir =
+    "./tweet-corpus/latvian-tweet-corpus-all-json-files-up-to-27-04-2021";
+  const files = fs.readdirSync(dir);
+
+  for (const file of files) {
+    const dateParts = file.split("-");
+    const rearrangedDateParts = [
+      dateParts[0],
+      dateParts[3],
+      dateParts[2],
+      dateParts[1],
+      ...dateParts.slice(4),
+    ];
+    rearrangedDateParts[2] =
+      rearrangedDateParts[2].length < 2
+        ? "0" + rearrangedDateParts[2]
+        : rearrangedDateParts[2];
+    rearrangedDateParts[3] =
+      rearrangedDateParts[3].length < 2
+        ? "0" + rearrangedDateParts[3]
+        : rearrangedDateParts[3];
+    const newFileName = rearrangedDateParts.join("-");
+    fs.rename(dir + "/" + file, dir + "/" + newFileName, (err) => {
+      if (err) console.log(err);
+    });
+  }
+};
+
 const readTweets = () => {
   // directory path
   const dir =
@@ -13,6 +54,11 @@ const readTweets = () => {
     }
 
     files.forEach((file) => {
+      console.log(
+        `Reading file ${file} (${files.findIndex((f) => f === file) + 1} of ${
+          files.length
+        })`
+      );
       let newTweets = fs.readFileSync(
         `./tweet-corpus/latvian-tweet-corpus-all-json-files-up-to-27-04-2021/${file}`,
         "utf-8"
@@ -25,7 +71,8 @@ const readTweets = () => {
               case brand.searchNames.some((name) =>
                 tweet.message.match(
                   new RegExp(
-                    ` (\\n)?(#)?${name}(\\n)? |^(\\n)?(#)?${name}|(#)?${name}(\\n)?$| (\\n)?@?${brand.accountName}(\\n)? |^(\\n)?@?${brand.accountName}|@?${brand.accountName}(\\n)?$`
+                    ` (\\n)?(#)?${name}(\\n)? |^(\\n)?(#)?${name}|(#)?${name}(\\n)?$| (\\n)?@?${brand.accountName}(\\n)? |^(\\n)?@?${brand.accountName}|@?${brand.accountName}(\\n)?$`,
+                    "ig"
                   )
                 )
               ):
@@ -34,7 +81,52 @@ const readTweets = () => {
                   "utf-8"
                 );
                 existingPeopleTweets = JSON.parse(existingPeopleTweets);
-                existingPeopleTweets.push(tweet);
+                let existingBrandTweetsToCheckRetweet = fs.readFileSync(
+                  brand.brandFile,
+                  "utf-8"
+                );
+                existingBrandTweetsToCheckRetweet = JSON.parse(
+                  existingBrandTweetsToCheckRetweet
+                );
+                //handle retweet
+                const originalTweetInPeopleTweets = returnOriginalTweet(
+                  tweet,
+                  existingPeopleTweets
+                );
+                if (originalTweetInPeopleTweets) {
+                  existingPeopleTweets = existingPeopleTweets.map(
+                    (existingTweet) => {
+                      if (
+                        existingTweet.tweetId ===
+                        originalTweetInPeopleTweets.tweetId
+                      )
+                        existingTweet.retweetCount++;
+                    }
+                  );
+                } else {
+                  const originalTweetInBrandTweets = returnOriginalTweet(
+                    tweet,
+                    existingBrandTweetsToCheckRetweet
+                  );
+                  if (originalTweetInBrandTweets) {
+                    existingBrandTweetsToCheckRetweet =
+                      existingBrandTweetsToCheckRetweet.map((existingTweet) => {
+                        if (
+                          existingTweet.tweetId ===
+                          originalTweetInBrandTweets.tweetId
+                        )
+                          existingTweet.retweetCount++;
+                      });
+                  } else {
+                    //check unique
+                    if (!tweetIsInArray(tweet, existingPeopleTweets)) {
+                      existingPeopleTweets.push({
+                        ...tweet,
+                        retweetCount: 0,
+                      });
+                    }
+                  }
+                }
                 const tweetsPeopleJson = JSON.stringify(existingPeopleTweets);
                 fs.writeFileSync(brand.peopleTweets, tweetsPeopleJson, "utf-8");
                 break;
@@ -45,9 +137,15 @@ const readTweets = () => {
                   "utf-8"
                 );
                 existingBrandTweets = JSON.parse(existingBrandTweets);
-                existingBrandTweets.push(tweet);
-                const tweetsBrandJson = JSON.stringify(existingBrandTweets);
-                fs.writeFileSync(brand.brandFile, tweetsBrandJson, "utf-8");
+                //check unique
+                if (
+                  !tweetIsInArray(tweet, existingBrandTweets) &&
+                  !tweet.retweetedId
+                ) {
+                  existingBrandTweets.push({ ...tweet, retweetCount: 0 });
+                  const tweetsBrandJson = JSON.stringify(existingBrandTweets);
+                  fs.writeFileSync(brand.brandFile, tweetsBrandJson, "utf-8");
+                }
                 break;
               default:
                 break;
@@ -59,7 +157,7 @@ const readTweets = () => {
   });
 };
 
-cleanFiles = () => {
+const cleanFiles = () => {
   // directory path
   const dir = "./brandTweets";
   const dir2 = "./peopleTweets";
